+++ backend/canisters/notifications_index/impl/src/model/subscriptions.rs
@@ -1,6 +1,7 @@
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::collections::hash_map::Entry::{Occupied, Vacant};
+use tracing::info;
 use types::{SubscriptionInfo, SubscriptionKeys, TimestampMillis, UserId};
 
 #[derive(Serialize, Deserialize, Default)]
@@ -95,6 +96,36 @@ impl Subscriptions {
 
     pub fn iter(&self) -> impl Iterator<Item = (&UserId, &Vec<SubscriptionInfoInternal>)> {
         self.subscriptions.iter()
+    }
+
+    // Return value is the vec of subscriptions removed per user.
+    // If the vec of subscriptions is empty, that means all of the user's subscriptions were removed.
+    pub fn remove_inactive(&mut self, cutoff: TimestampMillis) -> Vec<(UserId, Vec<String>)> {
+        let mut removed = Vec::new();
+        let mut count_removed = 0;
+
+        self.subscriptions.retain(|user_id, subscriptions| {
+            let removed_for_user: Vec<_> = subscriptions
+                .extract_if(.., |s| s.last_active < cutoff)
+                .map(|s| s.endpoint)
+                .collect();
+
+            count_removed += removed_for_user.len() as u32;
+
+            if subscriptions.is_empty() {
+                removed.push((*user_id, Vec::new()));
+                return false;
+            }
+
+            if !removed_for_user.is_empty() {
+                removed.push((*user_id, removed_for_user));
+            }
+            true
+        });
+
+        info!(count_removed, "Removed inactive subscriptions");
+
+        removed
     }
 }