+++ backend/libraries/group_chat_core/src/lib.rs
@@ -1061,7 +1061,7 @@ impl GroupChatCore {
 
     pub fn change_role(
         &mut self,
-        caller: UserId,
+        changed_by: UserId,
         target_users: Vec<UserId>,
         new_role: GroupRole,
         now: TimestampMillis,
@@ -1071,10 +1071,10 @@ impl GroupChatCore {
             bot_notification: None,
         };
 
+        let new_role_internal = new_role.into();
+
         for target_user in target_users {
-            let result = self
-                .members
-                .change_role(caller, target_user, new_role.into(), &self.permissions, now);
+            let result = self.members.change_role(target_user, new_role_internal, now);
 
             results.users.insert(target_user, result);
         }
@@ -1094,7 +1094,7 @@ impl GroupChatCore {
                 user_ids: successful_changes.iter().map(|(user_id, _)| *user_id).collect(),
                 old_role: successful_changes[0].1.into(),
                 new_role,
-                changed_by: caller,
+                changed_by,
             };
 
             let result = self
@@ -1765,6 +1765,7 @@ impl GroupChatCore {
         total_chit_earned: u32,
         streak: u16,
         streak_ends: TimestampMillis,
+        user_reauthenticated: bool,
     ) -> OCResult<ReservePrizeSuccess> {
         let member = self.members.get_verified_member(user_id)?;
         let min_visible_event_index = member.min_visible_event_index();
@@ -1779,6 +1780,7 @@ impl GroupChatCore {
             total_chit_earned,
             streak,
             streak_ends,
+            user_reauthenticated,
         )
     }