+++ backend/libraries/types/src/chat_summary.rs
@@ -10,7 +10,7 @@ use ts_export::ts_export;
 pub const MAX_THREADS_IN_SUMMARY: usize = 20;
 
 #[ts_export]
-#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+#[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct DirectChatSummary {
     pub them: UserId,
     pub last_updated: TimestampMillis,
@@ -31,7 +31,7 @@ pub struct DirectChatSummary {
 }
 
 #[ts_export]
-#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+#[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct DirectChatSummaryUpdates {
     pub chat_id: ChatId,
     pub last_updated: TimestampMillis,
@@ -82,7 +82,7 @@ pub struct PublicGroupSummary {
 }
 
 #[ts_export]
-#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+#[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct GroupCanisterGroupChatSummary {
     pub chat_id: ChatId,
     pub local_user_index_canister_id: CanisterId,
@@ -115,7 +115,7 @@ pub struct GroupCanisterGroupChatSummary {
 }
 
 #[ts_export]
-#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+#[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct GroupCanisterGroupChatSummaryUpdates {
     pub chat_id: ChatId,
     pub last_updated: TimestampMillis,
@@ -152,7 +152,7 @@ pub struct GroupCanisterGroupChatSummaryUpdates {
 }
 
 #[ts_export]
-#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+#[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct GroupMembership {
     pub joined: TimestampMillis,
     pub role: GroupRole,
@@ -166,7 +166,7 @@ pub struct GroupMembership {
 }
 
 #[ts_export]
-#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+#[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct GroupMembershipUpdates {
     pub role: Option<GroupRole>,
     pub mentions: Vec<HydratedMention>,
@@ -215,32 +215,65 @@ impl SelectedGroupUpdates {
 }
 
 #[ts_export]
-#[derive(CandidType, Serialize, Deserialize, Debug, Default, Clone)]
+#[derive(Serialize, Deserialize, Debug, Default, Clone)]
 pub struct ChatMetrics {
-    pub text_messages: u64,
-    pub image_messages: u64,
-    pub video_messages: u64,
-    pub audio_messages: u64,
-    pub file_messages: u64,
-    pub polls: u64,
-    pub poll_votes: u64,
-    pub crypto_messages: u64,
-    pub icp_messages: u64,
-    pub sns1_messages: u64,
-    pub ckbtc_messages: u64,
-    pub chat_messages: u64,
-    pub kinic_messages: u64,
-    pub deleted_messages: u64,
-    pub giphy_messages: u64,
-    pub prize_messages: u64,
-    pub prize_winner_messages: u64,
-    pub replies: u64,
-    pub edits: u64,
-    pub reactions: u64,
-    pub proposals: u64,
-    pub reported_messages: u64,
-    pub message_reminders: u64,
-    pub custom_type_messages: u64,
+    #[serde(default, skip_serializing_if = "ts_export::is_default")]
+    #[ts(as = "ts_export::TSBigIntWithDefault")]
+    pub text_messages: u32,
+    #[serde(default, skip_serializing_if = "ts_export::is_default")]
+    #[ts(as = "ts_export::TSBigIntWithDefault")]
+    pub image_messages: u32,
+    #[serde(default, skip_serializing_if = "ts_export::is_default")]
+    #[ts(as = "ts_export::TSBigIntWithDefault")]
+    pub video_messages: u32,
+    #[serde(default, skip_serializing_if = "ts_export::is_default")]
+    #[ts(as = "ts_export::TSBigIntWithDefault")]
+    pub audio_messages: u32,
+    #[serde(default, skip_serializing_if = "ts_export::is_default")]
+    #[ts(as = "ts_export::TSBigIntWithDefault")]
+    pub file_messages: u32,
+    #[serde(default, skip_serializing_if = "ts_export::is_default")]
+    #[ts(as = "ts_export::TSBigIntWithDefault")]
+    pub polls: u32,
+    #[serde(default, skip_serializing_if = "ts_export::is_default")]
+    #[ts(as = "ts_export::TSBigIntWithDefault")]
+    pub poll_votes: u32,
+    #[serde(default, skip_serializing_if = "ts_export::is_default")]
+    #[ts(as = "ts_export::TSBigIntWithDefault")]
+    pub crypto_messages: u32,
+    #[serde(default, skip_serializing_if = "ts_export::is_default")]
+    #[ts(as = "ts_export::TSBigIntWithDefault")]
+    pub deleted_messages: u32,
+    #[serde(default, skip_serializing_if = "ts_export::is_default")]
+    #[ts(as = "ts_export::TSBigIntWithDefault")]
+    pub giphy_messages: u32,
+    #[serde(default, skip_serializing_if = "ts_export::is_default")]
+    #[ts(as = "ts_export::TSBigIntWithDefault")]
+    pub prize_messages: u32,
+    #[serde(default, skip_serializing_if = "ts_export::is_default")]
+    #[ts(as = "ts_export::TSBigIntWithDefault")]
+    pub prize_winner_messages: u32,
+    #[serde(default, skip_serializing_if = "ts_export::is_default")]
+    #[ts(as = "ts_export::TSBigIntWithDefault")]
+    pub replies: u32,
+    #[serde(default, skip_serializing_if = "ts_export::is_default")]
+    #[ts(as = "ts_export::TSBigIntWithDefault")]
+    pub edits: u32,
+    #[serde(default, skip_serializing_if = "ts_export::is_default")]
+    #[ts(as = "ts_export::TSBigIntWithDefault")]
+    pub reactions: u32,
+    #[serde(default, skip_serializing_if = "ts_export::is_default")]
+    #[ts(as = "ts_export::TSBigIntWithDefault")]
+    pub proposals: u32,
+    #[serde(default, skip_serializing_if = "ts_export::is_default")]
+    #[ts(as = "ts_export::TSBigIntWithDefault")]
+    pub reported_messages: u32,
+    #[serde(default, skip_serializing_if = "ts_export::is_default")]
+    #[ts(as = "ts_export::TSBigIntWithDefault")]
+    pub message_reminders: u32,
+    #[serde(default, skip_serializing_if = "ts_export::is_default")]
+    #[ts(as = "ts_export::TSBigIntWithDefault")]
+    pub custom_type_messages: u32,
     pub last_active: TimestampMillis,
 }