+++ frontend/openchat-agent/src/services/community/community.client.ts
@@ -79,6 +79,7 @@ import {
     ResponseTooLargeError,
     isSuccessfulEventsResponse,
     offline,
+    random32,
     textToCode,
     toBigInt32,
 } from "openchat-shared";
@@ -211,6 +212,7 @@ import {
     apiMessageContent,
     apiUser as apiUserV2,
     apiVideoCallPresence,
+    changeRoleResult,
     createGroupSuccess,
     deletedMessageSuccess,
     enableOrResetInviteCodeSuccess,
@@ -329,27 +331,31 @@ export class CommunityClient extends MsgpackCanisterAgent {
         userId: string,
         newRole: MemberRole,
     ): Promise<ChangeRoleResponse> {
+        const user_id = principalStringToBytes(userId);
         return this.executeMsgpackUpdate(
             "change_channel_role",
             {
                 channel_id: toBigInt32(chatId.channelId),
-                user_id: principalStringToBytes(userId),
+                user_id,
+                user_ids: [user_id],
                 new_role: apiMemberRole(newRole),
             },
-            unitResult,
+            changeRoleResult,
             CommunityChangeChannelRoleArgs,
             UnitResult,
         );
     }
 
     changeRole(userId: string, newRole: MemberRole): Promise<ChangeCommunityRoleResponse> {
+        const user_id = principalStringToBytes(userId);
         return this.executeMsgpackUpdate(
             "change_role",
             {
-                user_id: principalStringToBytes(userId),
+                user_id,
+                user_ids: [user_id],
                 new_role: apiCommunityRole(newRole),
             },
-            unitResult,
+            changeRoleResult,
             CommunityChangeRoleArgs,
             UnitResult,
         );
@@ -367,7 +373,7 @@ export class CommunityClient extends MsgpackCanisterAgent {
                 history_visible_to_new_joiners: channel.historyVisible,
                 avatar: mapOptional(channel.avatar?.blobData, (data) => {
                     return {
-                        id: DataClient.newBlobId(),
+                        id: BigInt(random32()),
                         data,
                         mime_type: "image/jpg",
                     };
@@ -1399,7 +1405,7 @@ export class CommunityClient extends MsgpackCanisterAgent {
                         ? "NoChange"
                         : {
                               SetToSome: {
-                                  id: DataClient.newBlobId(),
+                                  id: BigInt(random32()),
                                   mime_type: "image/jpg",
                                   data: avatar,
                               },
@@ -1445,7 +1451,7 @@ export class CommunityClient extends MsgpackCanisterAgent {
                         ? "NoChange"
                         : {
                               SetToSome: {
-                                  id: DataClient.newBlobId(),
+                                  id: BigInt(random32()),
                                   mime_type: "image/jpg",
                                   data: avatar,
                               },
@@ -1455,7 +1461,7 @@ export class CommunityClient extends MsgpackCanisterAgent {
                         ? "NoChange"
                         : {
                               SetToSome: {
-                                  id: DataClient.newBlobId(),
+                                  id: BigInt(random32()),
                                   mime_type: "image/jpg",
                                   data: banner,
                               },