+++ frontend/openchat-agent/src/utils/caching.ts
@@ -57,7 +57,7 @@ import {
     updateCreatedUser,
 } from "openchat-shared";
 
-const CACHE_VERSION = 144;
+const CACHE_VERSION = 145;
 const EARLIEST_SUPPORTED_MIGRATION = 138;
 const MAX_INDEX = 9999999999;
 
@@ -268,6 +268,7 @@ const migrations: Record<number, MigrationFunction<ChatSchema>> = {
     142: createPublicProfileStore,
     143: clearCommunityDetailsStore,
     144: clearChatsStore,
+    145: clearChatsStore,
 };
 
 async function migrate(
@@ -412,10 +413,7 @@ export async function getCachedChats(
     const resolvedDb = await db;
     const chats = await resolvedDb.get("chats", principal.toString());
 
-    if (
-        chats !== undefined &&
-        chats.latestUserCanisterUpdates < BigInt(Date.now() - 30 * ONE_DAY)
-    ) {
+    if (chats && chats.latestUserCanisterUpdates < BigInt(Date.now() - 30 * ONE_DAY)) {
         // If the cache was last updated more than 30 days ago, clear the cache and return undefined
         const storeNames = resolvedDb.objectStoreNames;
         for (let i = 0; i < storeNames.length; i++) {