+++ frontend/openchat-shared/src/domain/chat/chat.ts
@@ -2209,9 +2209,7 @@ export type MessageAction = "emoji" | "file" | undefined;
 export type Metrics = {
     audioMessages: number;
     edits: number;
-    icpMessages: number;
-    sns1Messages: number;
-    ckbtcMessages: number;
+    cryptoMessages: number;
     giphyMessages: number;
     deletedMessages: number;
     reportedMessages: number;